+++ lib/slimi/railtie.rb
@@ -5,6 +5,7 @@ module Slimi
     initializer 'Register Slimi template handler' do
       ::ActiveSupport.on_load(:action_view) do
         ::ActionView::Template.register_template_handler(
+          :slim,
           RailsTemplateHandler.new
         )
       end